Contract acquisition, storage of combinator.
Added function and test for contract acquisition. Implemented correct storage of contract combinators using the pwasm Box (should work when deployed, easier than calling write/read manually). Likely will need to fill in combinators by using financial_sc.rs as a template.

+++ financial-smart-contract/src/combinators/and_combinator.rs
@@ -21,11 +21,11 @@ impl<'a, T: ContractCombinator, U: ContractCombinator> AndCombinator<'a, T, U> {
 // Contract combinator implementation for the and combinator
 impl<'a, T: ContractCombinator, U: ContractCombinator> ContractCombinator for AndCombinator<'a, T, U> {
     // Returns the latest of the two sub-horizons
-    fn get_horizon(&self) -> Option<i32> {
+    fn get_horizon(&self) -> Option<u64> {
         latest_horizon(self.sub_combinator0.get_horizon(), self.sub_combinator1.get_horizon())
     }
 
-    fn acquire(&self, time: i32) -> i32 {
+    fn acquire(&self, time: u64) -> u64 {
         self.sub_combinator0.acquire(time) + self.sub_combinator1.acquire(time)
     }
 }
