+++ financial-smart-contract/src/combinators/contract_combinator.rs
@@ -1,12 +1,12 @@
 // API for combinators
 pub trait ContractCombinator {
     // Returns the horizon of the combinator, or -1 if none exists
-    fn get_horizon(&self) -> Option<i32> {
+    fn get_horizon(&self) -> Option<u64> {
         None
     }
 
     // Returns whether the given time is beyond the combinator's horizon or not
-    fn past_horizon(&self, time: i32) -> bool {
+    fn past_horizon(&self, time: u64) -> bool {
         match self.get_horizon() {
             Some(t) => t < time,
             None => false,
@@ -14,11 +14,11 @@ pub trait ContractCombinator {
     }
 
     // Returns the value of the combinator when acquired at the given time
-    fn acquire(&self, time: i32) -> i32;
+    fn acquire(&self, time: u64) -> u64;
 }
 
 // Returns the earliest of the given horizons
-pub fn earliest_horizon(horizon0: Option<i32>, horizon1: Option<i32>) -> Option<i32> {
+pub fn earliest_horizon(horizon0: Option<u64>, horizon1: Option<u64>) -> Option<u64> {
     match horizon0 {
         Some(h0) => match horizon1 {
             Some(h1) => if h0 < h1 {
@@ -33,7 +33,7 @@ pub fn earliest_horizon(horizon0: Option<i32>, horizon1: Option<i32>) -> Option<
 }
 
 // Returns the latest of the given horizons
-pub fn latest_horizon(horizon0: Option<i32>, horizon1: Option<i32>) -> Option<i32> {
+pub fn latest_horizon(horizon0: Option<u64>, horizon1: Option<u64>) -> Option<u64> {
     match horizon0 {
         Some(h0) => match horizon1 {
             Some(h1) => if h0 > h1 {
